+++ src/operator/tensor/cast_storage-inl.h
@@ -3,8 +3,8 @@
  * \file cast_storage-inl.h
  * \brief cast_storage implementation for dense and sparse tensors
  */
-#ifndef MXNET_OPERATOR_NN_CAST_STORAGE_INL_H_
-#define MXNET_OPERATOR_NN_CAST_STORAGE_INL_H_
+#ifndef MXNET_OPERATOR_TENSOR_CAST_STORAGE_INL_H_
+#define MXNET_OPERATOR_TENSOR_CAST_STORAGE_INL_H_
 
 #include <dmlc/timer.h>
 #include <mxnet/ndarray.h>
@@ -333,4 +333,4 @@ void CastStorageComputeEx(const nnvm::NodeAttrs& attrs,
 }  // namespace op
 }  // namespace mxnet
 
-#endif  // MXNET_OPERATOR_NN_CAST_STORAGE_INL_H_
+#endif  // MXNET_OPERATOR_TENSOR_CAST_STORAGE_INL_H_
